+++ src/vatensor/vmath.h
@@ -16,6 +16,7 @@ namespace va {
 
     void sign(VArrayTarget target, const VArray& array);
     void abs(VArrayTarget target, const VArray& array);
+    void square(VArrayTarget target, const VArray& array);
     void sqrt(VArrayTarget target, const VArray& array);
     void exp(VArrayTarget target, const VArray& array);
     void log(VArrayTarget target, const VArray& array);